requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
ulid-py>=1.1,<2.0
